Adding new doc menu to sidebar items
